+++ popcor/base_lifters/__init__.py
@@ -1,4 +1,5 @@
-from .poly_lifters import PolyLifter
+from .poly_lifter import PolyLifter
+from .range_only_lifter import RangeOnlyLifter
 from .robust_pose_lifter import RobustPoseLifter
 from .state_lifter import StateLifter
 from .stereo_lifter import StereoLifter