+++ popcor/examples/__init__.py
@@ -2,9 +2,24 @@ from .example_lifter import ExampleLifter
 from .mono_lifter import MonoLifter
 from .poly4_lifter import Poly4Lifter
 from .poly6_lifter import Poly6Lifter
-from .range_only_lifters import RangeOnlyLocLifter
+from .ro_nsq_lifter import RangeOnlyNsqLifter
+from .ro_sq_lifter import RangeOnlySqLifter
 from .rotation_lifter import RotationLifter
 from .stereo1d_lifter import Stereo1DLifter
 from .stereo2d_lifter import Stereo2DLifter
 from .stereo3d_lifter import Stereo3DLifter
 from .wahba_lifter import WahbaLifter
+
+__all__ = [  # type: ignore
+    ExampleLifter,
+    MonoLifter,
+    Poly4Lifter,
+    Poly6Lifter,
+    RangeOnlyNsqLifter,
+    RangeOnlySqLifter,
+    RotationLifter,
+    Stereo1DLifter,
+    Stereo2DLifter,
+    Stereo3DLifter,
+    WahbaLifter,
+]